split store implementation
