+++ fiber-bin/Cargo.toml
@@ -1,6 +1,22 @@
 [package]
 name = "fiber-bin"
-version = "0.1.0"
+version = "0.4.2"
 edition = "2021"
 
 [dependencies]
+fnn = { path = "../fiber-lib" }
+ckb-chain-spec    = "0.200.0"
+ckb-resource      = "0.200.0"
+ractor = "0.14.2"
+tokio = { version = "1", features = [
+    "io-util",
+    "macros",
+    "rt",
+    "rt-multi-thread",
+    "sync",
+    "net",
+    "time",
+    "signal",
+] }
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
