+++ fiber-lib/src/store/mod.rs
@@ -1,7 +1,7 @@
 pub mod db_migrate;
 pub mod migration;
 mod schema;
-pub mod store;
-pub use store::Store;
+pub mod store_impl;
+pub use store_impl::Store;
 #[cfg(test)]
 mod tests;
