+++ fiber-lib/src/store/tests/store.rs
@@ -15,8 +15,8 @@ use crate::gen_rand_fiber_public_key;
 use crate::gen_rand_sha256_hash;
 use crate::invoice::*;
 use crate::now_timestamp_as_millis_u64;
-use crate::store::store::deserialize_from;
-use crate::store::store::serialize_to_vec;
+use crate::store::store_impl::deserialize_from;
+use crate::store::store_impl::serialize_to_vec;
 use crate::store::Store;
 use crate::watchtower::*;
 use ckb_hash::new_blake2b;
